Changes to be committed: 	modified:   port/boards/labplus_Ledong_pro/modules/mpython.py 	modified:   port/modules/neopixel.py

+++ port/modules/neopixel.py
@@ -43,7 +43,7 @@ class NeoPixel:
             c = v[i]
             j = self.ORDER[i]
             while j < l:
-                b[j] = c * self._brightness
+                b[j] = int(c * self._brightness)
                 j += bpp
 
     def write(self):
